Interceptor

+++ server/src/post/post.service.ts
@@ -6,29 +6,64 @@ import {
 import { CreatePostDto } from './dto/create-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from './post.entity';
-import { Repository } from 'typeorm';
+import { Brackets, Repository } from 'typeorm';
+import { User } from 'src/auth/user.entity';
+import { Image } from 'src/image/image.entity';
 
-// 서비스: 각 컨트롤러의 메서드에 해당하는 DB 로직을 구현함.
 @Injectable()
 export class PostService {
   constructor(
     @InjectRepository(Post) // 해당 엔티티를 가진 레포지토리를 주입함.
     private postRepository: Repository<Post>,
+
+    @InjectRepository(Image)
+    private imageRepository: Repository<Image>,
   ) {}
 
-  async getPosts(page: number) {
+  private getPostsWithOrderImages(posts: Post[]) {
+    return posts.map((post) => {
+      const { images, ...rest } = post;
+      const newImages = [...images].sort((a, b) => a.id - b.id);
+      return { ...rest, images: newImages };
+    });
+  }
+
+  private async getPostsBaseQuery(
+    userId: number,
+  ): Promise<SelectQueryBuilder<Post>> {
+    return this.postRepository
+      .createQueryBuilder('post')
+      .leftJoinAndSelect('post.images', 'image')
+      .where('post.userId = :userId', { userId })
+      .orderBy('post.date', 'DESC');
+  }
+
+  async getMyPosts(page: number, user: User) {
+    const perPage = 10;
+    const offset = (page - 1) * perPage;
+    const queryBuilder = await this.getPostsBaseQuery(user.id);
+    const posts = await queryBuilder.take(perPage).skip(offset).getMany();
+
+    return this.getPostsWithOrderImages(posts);
+  }
+
+  async getPosts(page: number, user: User) {
     const perPage = 10;
     const offset = (page - 1) * perPage;
     // 레포지토리.createQueryBuilder('테이블명'): 해당 테이블에서 데이터를 가져옴.
-    return this.postRepository
+    const posts = await this.postRepository
       .createQueryBuilder('post')
+      .leftJoinAndSelect('post.images', 'image')
+      .where('post.userId = :userId', { userId: user.id })
       .orderBy('post.date', 'DESC')
       .take(perPage) // 페이지 당 개수
       .skip(offset) // 건너뛸 페이지 개수
       .getMany(); // 여러 개를 가져옴.
+
+    return this.getPostsWithOrderImages(posts);
   }
 
-  async createPost(createPostDto: CreatePostDto) {
+  async createPost(createPostDto: CreatePostDto, user: User) {
     const [
       latitude,
       longitude,
@@ -52,9 +87,14 @@ export class PostService {
       date,
       score,
       imageUris,
+      user,
     });
 
+    const images = imageUris.map((uri) => this.imageRepository.create(uri));
+    post.images = images;
+
     try {
+      await this.imageRepository.save(images);
       await this.postRepository.save(post); // 레포지토리.save(데이터): 데이터를 실제 DB에 저장함.
     } catch (error) {
       console.log(error);
@@ -63,20 +103,33 @@ export class PostService {
       );
     }
 
-    return post;
+    const { user: _, ...postWithoutUser } = post;
+    return postWithoutUser;
   }
 
-  async getPostById(id: number) {
+  async getPostById(id: number, user: User) {
     try {
       const foundPost = await this.postRepository
         .createQueryBuilder('post')
-        .where('post.id = :id', { id }) // 매개변수로 받은 id를 조건식으로 넘겨줌.
+        .leftJoinAndSelect('post.images', 'image')
+        .leftJoinAndSelect(
+          'post.favorites',
+          'favorite',
+          'favorite.userId = :userId',
+          { userId: user.id },
+        )
+        .where('post.userId = :userId', { userId: user.id }) // 조건을 추가함.
+        .andWhere('post.id = :id', { id }) // 매개변수로 받은 id를 조건식으로 넘겨줌.
         .getOne(); // 데이터를 한 개만 가져옴.
 
       if (!foundPost) {
         throw new NotFoundException('존재하지 않는 피드입니다.');
       }
-      return foundPost;
+
+      const { favorites, ...rest } = foundPost;
+      const postWithIsFavorites = { ...rest, isFavorite: favorites.length > 0 };
+
+      return postWithIsFavorites;
     } catch (error) {
       console.log(error);
       throw new InternalServerErrorException(
@@ -85,13 +138,14 @@ export class PostService {
     }
   }
 
-  async deletePost(id: number) {
+  async deletePost(id: number, user: User) {
     try {
       const result = await this.postRepository
         .createQueryBuilder('post')
         .delete()
         .from(Post)
-        .where('id = :id', { id })
+        .where('userId = :userId', { userId: user.id })
+        .andWhere('id = :id', { id })
         .execute(); // 해당 작업(삭제)을 실행함.
 
       if (result.affected === 0) {
@@ -111,8 +165,9 @@ export class PostService {
   async updatePost(
     id: number,
     updatePostDto: Omit<CreatePostDto, 'latitude' | 'longitude' | 'address'>,
+    user: User,
   ) {
-    const post = await this.getPostById(id);
+    const post = await this.getPostById(id, user);
 
     const { title, description, color, date, score, imageUris } = updatePostDto;
 
@@ -133,10 +188,11 @@ export class PostService {
     return post;
   }
 
-  async getAllMarkers() {
+  async getAllMarkers(user: User) {
     try {
       const markers = await this.postRepository
         .createQueryBuilder('post')
+        .where('post.userId = :userId', { userId: user.id })
         .select([
           // 해당 컬럼의 값들만 선택하여 가져옴.
           'post.id',
@@ -154,5 +210,55 @@ export class PostService {
         '마커를 불러오는 도중 에러가 발생했습니다.',
       );
     }
+  }
+
+  async getPostsByMonth(year: number, month: number, user: User) {
+    const posts = await this.postRepository
+      .createQueryBuilder('post')
+      .where('post.userId = :userId', { userId: user.id })
+      .andWhere('extract(year from post.date) = :year', { year })
+      .andWhere('extract(month from post.date) = :month', { month })
+      .select([
+        'post.id AS id',
+        'post.title AS title',
+        'post.address AS address',
+        'EXTRACT(DAY FROM post.date) AS date',
+      ])
+      .getRawMany();
+
+    const groupPostsByDate = posts.reduce((acc, post) => {
+      const { id, title, address, date } = post;
+
+      if (!acc[date]) {
+        acc[date] = [];
+      }
+      acc[date].push({ id, title, address });
+
+      return acc;
+    }, {});
+
+    return groupPostsByDate;
+  }
+
+  async searchMyPostsByTitleAndAddress(
+    query: string,
+    page: number,
+    user: User,
+  ) {
+    const perPage = 10;
+    const offset = (page - 1) * perPage;
+    const queryBuilder = await this.getPostsBaseQuery(user.id);
+    const posts = await queryBuilder
+      .andWhere(
+        new Brackets((qb) => {
+          qb.where('post.title like: query', { query: `%${query}%` });
+          qb.orWhere('post.address like:query', { query: `%${query}%` });
+        }),
+      )
+      .skip(offset)
+      .take(perPage)
+      .getMany();
+
+    return this.getPostsWithOrderImages(posts);
   }
 }
